requirements.txt
streamlit>=1.40.0
requests>=2.32.0
python-dotenv>=1.0.1
edge-tts>=6.1.0
yt-dlp>=2025.1.26
